chore: type, export, 파일 분리 컨벤션 적용
