+++ frontend/src/@common/components/Header/Header.types.ts
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react';
+
+interface HeaderProps extends PropsWithChildren {
+  isHome?: boolean;
+}
+
+export type { HeaderProps };
