+++ frontend/src/@common/components/HomeButton/HomeButton.types.ts
@@ -1,0 +1,8 @@
+interface HomeButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  icon: string;
+  isHome?: boolean;
+  onClick: () => void;
+}
+
+export type { HomeButtonProps };
