+++ frontend/src/@common/components/ToggleSwitch/ToggleSwitch.types.ts
@@ -1,0 +1,6 @@
+interface ToggleSwitchProps {
+  active: boolean;
+  onToggle: () => void;
+}
+
+export type { ToggleSwitchProps };
