+++ frontend/src/@common/contexts/ToastProvider.tsx
@@ -1,0 +1,14 @@
+import { useToast } from '../hooks/useToast';
+import { ToastContext } from './useToastContext';
+
+const ToastProvider = ({ children }: { children: React.ReactNode }) => {
+  const { toast, showToast } = useToast();
+
+  return (
+    <ToastContext.Provider value={{ toast, showToast }}>
+      {children}
+    </ToastContext.Provider>
+  );
+};
+
+export default ToastProvider;
