+++ frontend/src/@common/contexts/useToastContext.ts
@@ -1,7 +1,5 @@
 import { createContext, useContext } from 'react';
 
-import { useToast } from '../hooks/useToast';
-
 import type { ToastContextProps } from '../components/Toast/Toast.types';
 
 const ToastContext = createContext<ToastContextProps>({
@@ -9,17 +7,7 @@ const ToastContext = createContext<ToastContextProps>({
   showToast: () => {},
 });
 
-export const ToastProvider = ({ children }: { children: React.ReactNode }) => {
-  const { toast, showToast } = useToast();
-
-  return (
-    <ToastContext.Provider value={{ toast, showToast }}>
-      {children}
-    </ToastContext.Provider>
-  );
-};
-
-export const useToastContext = () => {
+const useToastContext = () => {
   const context = useContext(ToastContext);
 
   if (!context) {
@@ -28,3 +16,5 @@ export const useToastContext = () => {
 
   return context;
 };
+
+export { ToastContext, useToastContext };
